fix: move desktopbridge dependency to website
